backend/requirements.txt
airtable-python-wrapper==0.15.3
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.2
billiard==4.1.0
Brotli==1.1.0
cached-property==1.5.2
cachetools==5.5.0
celery==5.3.1
certifi==2024.2.2
cffi==1.16.0
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.3.2
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coinbase-commerce==1.0.1
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.0
cryptography==43.0.3
daphne==4.1.2
defusedxml==0.7.1
dj-database-url==2.3.0
dj-rest-auth==6.0.0
Django==4.2.7
django-4-jet==1.0.9
django-admin==2.0.2
django-allauth==65.0.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-countries==7.5.1
django-encrypted-fields==1.1.2
django-encrypted-model-fields==0.6.5
django-excel-response2==3.0.6
django-filter==24.3
django-ordered-model==3.7.4
django-rest-framework==0.1.0
django-six==1.0.5
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dlib==19.24.6
dnspython==2.4.2
emoji==2.8.0
excel-base==1.0.4
exceptiongroup==1.2.2
exchangelib==5.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fake-useragent==2.0.2
fastapi==0.104.1
Flask==2.2.2
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gunicorn==21.2.0
h11==0.13.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.6.1
importlib_resources==6.4.5
incremental==24.7.2
iniconfig==2.0.0
instagrapi==2.1.3
isodate==0.6.1
isoweek==1.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.1
mongoengine==0.27.0
motor==3.3.2
msgpack==1.1.0
mutagen==1.47.0
numpy==2.0.2
oauth2==1.9.0.post1
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
outcome==1.2.0
packaging==23.1
phonenumbers==9.0.1
Pillow==9.3.0
pluggy==1.3.0
prompt-toolkit==3.0.38
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.17.2
PyJWT==2.7.0
pymongo==4.6.0
pyOpenSSL==25.0.0
pyparsing==3.2.0
PySocks==1.7.1
pyspnego==0.10.2
pytesseract==0.3.13
pytest==7.4.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.20.0
python-keyczar==0.716
python-multipart==0.0.6
python3-openid==3.2.0
pytz==2023.3
PyVirtualDisplay==3.0
PyYAML==6.0.2
redis==4.6.0
requests==2.31.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
rsa==4.9
screen==1.0.1
selenium==4.4.3
service-identity==24.2.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
starlette==0.27.0
stripe==8.9.0
TimeConvert==3.0.13
tomli==2.0.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
tweepy==4.15.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
undetected-chromedriver==3.1.5.post4
uritemplate==4.1.1
urllib3==1.26.19
uvicorn==0.24.0.post1
vine==5.0.0
wcwidth==0.2.6
webdriver-manager==3.8.3
websockets==10.3
Werkzeug==2.2.2
whitenoise==6.5.0
wsproto==1.1.0
xlwt==1.3.0
zipp==3.21.0
zope.interface==7.2
